requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
torch>=2.0
lightgbm>=4.0
dbfread>=2.0
pysus>=0.10
fastapi>=0.104
uvicorn[standard]>=0.24
# Platform Flask (produção: gunicorn -w 4 -b 0.0.0.0:5001 "run:app")
gunicorn>=21.0
flask>=3.0
flask-sqlalchemy>=3.1
flask-login>=0.6
flask-jwt-extended>=4.6
flask-cors>=4.0
pymysql>=1.1
python-dotenv>=1.0
bcrypt>=4.0
pyotp>=2.9
qrcode[pil]>=7.4
